Introduce run node command
