+++ waspc/src/Wasp/Generator/DockerGenerator.hs
@@ -19,7 +19,6 @@ import qualified Wasp.AppSpec.Entity as AS.Entity
 import Wasp.Generator.Common
   ( ProjectRootDir,
     ServerRootDir,
-    latestMajorNodeVersion,
   )
 import Wasp.Generator.DbGenerator.Common
   ( PrismaDbSchema,
@@ -32,6 +31,7 @@ import Wasp.Generator.Monad (Generator, GeneratorError, runGenerator)
 import Wasp.Generator.Templates (TemplatesDir, compileAndRenderTemplate)
 import qualified Wasp.SemanticVersion as SV
 import Wasp.Util (getEnvVarDefinition)
+import Wasp.Util.NodeVersion (latestMajorNodeVersion)
 
 genDockerFiles :: AppSpec -> Generator [FileDraft]
 genDockerFiles spec = sequence [genDockerfile spec, genDockerignore spec]
