+++ waspc/src/Wasp/Generator/WebAppGenerator.hs
@@ -29,7 +29,6 @@ import Wasp.AppSpec.Valid (getApp, isAuthEnabled)
 import Wasp.Env (envVarsToDotEnvContent)
 import Wasp.Generator.Common
   ( makeJsonWithEntityData,
-    nodeVersionRange,
     prismaVersion,
   )
 import qualified Wasp.Generator.ConfigFile as G.CF
@@ -48,6 +47,7 @@ import Wasp.Generator.WebAppGenerator.OperationsGenerator (genOperations)
 import Wasp.Generator.WebAppGenerator.RouterGenerator (genRouter)
 import qualified Wasp.SemanticVersion as SV
 import Wasp.Util ((<++>))
+import qualified Wasp.Util.NodeVersion as NodeVersion
 
 genWebApp :: AppSpec -> Generator [FileDraft]
 genWebApp spec = do
@@ -101,7 +101,7 @@ genPackageJson spec waspDependencies = do
             [ "appName" .= (fst (getApp spec) :: String),
               "depsChunk" .= N.getDependenciesPackageJsonEntry combinedDependencies,
               "devDepsChunk" .= N.getDevDependenciesPackageJsonEntry combinedDependencies,
-              "nodeVersionRange" .= show nodeVersionRange
+              "nodeVersionRange" .= show NodeVersion.nodeVersionRange
             ]
       )
