+++ waspc/waspc.cabal
@@ -308,6 +308,8 @@ library
     Wasp.Util.Fib
     Wasp.Util.IO
     Wasp.Util.Terminal
+    Wasp.Util.NodeVersion
+    Wasp.Util.NodeCommand
     Wasp.WaspignoreFile
     Wasp.Generator.NpmDependencies
     Wasp.Generator.NpmInstall
